+++ packages/next-intl/src/plugin/createMessagesDeclaration.tsx
