+++ packages/next-intl/src/plugin/createNextIntlPlugin.tsx
@@ -1,5 +1,5 @@
 import type {NextConfig} from 'next';
-import createMessagesDeclarations from './createMessagesDeclarations.js';
+import createMessagesDeclaration from './createMessagesDeclaration.js';
 import getNextConfig from './getNextConfig.js';
 import type {PluginConfig} from './types.js';
 import {warn} from './utils.js';
